+++ Cpp/test/sir_bernoulli.cpp
@@ -7,10 +7,10 @@
 
 TEST(SIR_Bernoulli_Network_Test, run_simulation_test)
 {
-    uint16_t N_pop = 100;
+    uint32_t N_pop = 100;
     float p_ER = .4;
     float p_I0 = .1;
-    uint16_t N_steps = 100;
+    uint32_t N_steps = 100;
 
     std::random_device rd;
     std::mt19937 generator(rd());
@@ -19,13 +19,13 @@ TEST(SIR_Bernoulli_Network_Test, run_simulation_test)
 
     Network_Models::SIR_Bernoulli_Network G(N_pop, p_ER, generator);
     G.generate_initial_infections(p_I0, p_R0);
-    uint16_t Nt = 100;
+    uint32_t Nt = 100;
     float p_R = 0.01;
     float p_I = 0.1;
     auto traj = G.simulate(Nt, p_I, p_R);
 
     std::ofstream outfile("test_traj.txt");
-    for (uint16_t i = 0; i < traj.size(); i++) {
+    for (uint32_t i = 0; i < traj.size(); i++) {
         for (auto val: traj[i]) {
             outfile << val << ",";
         }
